+++ src/core/expressions/baseExpression.ts
@@ -640,407 +640,4 @@ module Core {
     }
   }
   check = Expression;
-
-
-
-// =====================================================================================
-// =====================================================================================
-
-  export class UnaryExpression extends Expression {
-    static jsToValue(parameters: ExpressionJS): ExpressionValue {
-      var value: ExpressionValue = {
-        op: parameters.op
-      };
-      if (parameters.operand) {
-        value.operand = Expression.fromJSLoose(parameters.operand);
-      } else {
-        throw new TypeError("must have a operand");
-      }
-
-      return value;
-    }
-
-    public operand: Expression;
-
-    constructor(parameters: ExpressionValue, dummyObject: Dummy) {
-      super(parameters, dummyObject);
-      this.operand = parameters.operand;
-    }
-
-    public valueOf(): ExpressionValue {
-      var value = super.valueOf();
-      value.operand = this.operand;
-      return value;
-    }
-
-    public toJS(): ExpressionJS {
-      var js: ExpressionJS = super.toJS();
-      js.operand = this.operand.toJS();
-      return js;
-    }
-
-    public equals(other: UnaryExpression): boolean {
-      return super.equals(other) &&
-        this.operand.equals(other.operand)
-    }
-
-    public getComplexity(): number {
-      return 1 + this.operand.getComplexity()
-    }
-
-    protected _specialSimplify(simpleOperand: Expression): Expression {
-      return null;
-    }
-
-    public simplify(): Expression {
-      var simpleOperand = this.operand.simplify();
-
-      var special = this._specialSimplify(simpleOperand);
-      if (special) return special;
-
-      if (simpleOperand.isOp('literal')) {
-        return new LiteralExpression({
-          op: 'literal',
-          value: this._makeFn(simpleOperand.getFn())()
-        })
-      }
-
-      var value = this.valueOf();
-      value.operand = simpleOperand;
-      return new (Expression.classMap[this.op])(value);
-    }
-
-    public getReferences(): string[] {
-      return this.operand.getReferences();
-    }
-
-    public getOperandOfType(type: string): Expression[] {
-      if (this.operand.isOp(type)) {
-        return [this.operand];
-      } else {
-        return []
-      }
-    }
-
-    public substitute(substitutionFn: SubstitutionFn, genDiff: number): Expression {
-      var sub = substitutionFn(this, genDiff);
-      if (sub) return sub;
-      var subOperand = this.operand.substitute(substitutionFn, genDiff);
-      if (this.operand === subOperand) return this;
-      var value = this.valueOf();
-      value.operand = subOperand;
-      return new (Expression.classMap[this.op])(value);
-    }
-
-    protected _makeFn(operandFn: Function): Function {
-      throw new Error("should never be called directly");
-    }
-
-    public getFn(): Function {
-      return this._makeFn(this.operand.getFn());
-    }
-
-    protected _makeFnJS(operandFnJS: string): string {
-      throw new Error("should never be called directly");
-    }
-
-    /* protected */
-    public _getRawFnJS(): string {
-      return this._makeFnJS(this.operand._getRawFnJS())
-    }
-
-    protected _checkTypeOfOperand(wantedType: string): void {
-      if (!this.operand.canHaveType(wantedType)) {
-        throw new TypeError(this.op + ' expression must have an operand of type ' + wantedType);
-      }
-    }
-
-    public _fillRefSubstitutions(typeContext: any, alterations: Alteration[]): any {
-      this.operand._fillRefSubstitutions(typeContext, alterations);
-      return this.type;
-    }
-  }
-
-// =====================================================================================
-// =====================================================================================
-
-  export class BinaryExpression extends Expression {
-    static jsToValue(parameters: ExpressionJS): ExpressionValue {
-      var op = parameters.op;
-      var value: ExpressionValue = {
-        op: op
-      };
-      if (typeof parameters.lhs !== 'undefined' && parameters.lhs !== null) {
-        value.lhs = Expression.fromJSLoose(parameters.lhs);
-      } else {
-        throw new TypeError("must have a lhs");
-      }
-
-      if (typeof parameters.rhs !== 'undefined' && parameters.rhs !== null) {
-        value.rhs = Expression.fromJSLoose(parameters.rhs);
-      } else {
-        throw new TypeError("must have a rhs");
-      }
-
-      return value;
-    }
-
-    public lhs: Expression;
-    public rhs: Expression;
-
-    protected simple: boolean;
-
-    constructor(parameters: ExpressionValue, dummyObject: Dummy) {
-      super(parameters, dummyObject);
-      this.lhs = parameters.lhs;
-      this.rhs = parameters.rhs;
-    }
-
-    public valueOf(): ExpressionValue {
-      var value = super.valueOf();
-      value.lhs = this.lhs;
-      value.rhs = this.rhs;
-      return value;
-    }
-
-    public toJS(): ExpressionJS {
-      var js = super.toJS();
-      js.lhs = this.lhs.toJS();
-      js.rhs = this.rhs.toJS();
-      return js;
-    }
-
-    public equals(other: BinaryExpression): boolean {
-      return super.equals(other) &&
-        this.lhs.equals(other.lhs) &&
-        this.rhs.equals(other.rhs);
-    }
-
-    public getComplexity(): number {
-      return 1 + this.lhs.getComplexity() + this.rhs.getComplexity()
-    }
-
-    protected _specialSimplify(simpleLhs: Expression, simpleRhs: Expression): Expression {
-      return null;
-    }
-
-    public simplify(): Expression {
-      var simpleLhs = this.lhs.simplify();
-      var simpleRhs = this.rhs.simplify();
-
-      var special = this._specialSimplify(simpleLhs, simpleRhs);
-      if (special) return special;
-
-      if (simpleLhs.isOp('literal') && simpleRhs.isOp('literal')) {
-        return new LiteralExpression({
-          op: 'literal',
-          value: this._makeFn(simpleLhs.getFn(), simpleRhs.getFn())()
-        })
-      }
-
-      var value = this.valueOf();
-      value.lhs = simpleLhs;
-      value.rhs = simpleRhs;
-      return new (Expression.classMap[this.op])(value);
-    }
-
-    public getOperandOfType(type: string): Expression[] {
-      var ret: Expression[] = [];
-
-      if (this.lhs.isOp(type)) ret.push(this.lhs);
-      if (this.rhs.isOp(type)) ret.push(this.rhs);
-      return ret;
-    }
-
-    public checkLefthandedness(): boolean {
-      if (this.lhs instanceof RefExpression && this.rhs instanceof RefExpression) return null;
-      if (this.lhs instanceof RefExpression) return true;
-      if (this.rhs instanceof RefExpression) return false;
-
-      return null;
-    }
-
-    public getReferences(): string[] {
-      return this.lhs.getReferences().concat(this.rhs.getReferences()).sort();
-    }
-
-    public substitute(substitutionFn: SubstitutionFn, genDiff: number): Expression {
-      var sub = substitutionFn(this, genDiff);
-      if (sub) return sub;
-      var subLhs = this.lhs.substitute(substitutionFn, genDiff);
-      var subRhs = this.rhs.substitute(substitutionFn, genDiff);
-      if (this.lhs === subLhs && this.rhs === subRhs) return this;
-      var value = this.valueOf();
-      value.lhs = subLhs;
-      value.rhs = subRhs;
-      return new (Expression.classMap[this.op])(value);
-    }
-
-    protected _makeFn(lhsFn: Function, rhsFn: Function): Function {
-      throw new Error("should never be called directly");
-    }
-
-    public getFn(): Function {
-      return this._makeFn(this.lhs.getFn(), this.rhs.getFn());
-    }
-
-    protected _makeFnJS(lhsFnJS: string, rhsFnJS: string): string {
-      throw new Error("should never be called directly");
-    }
-
-    /* protected */
-    public _getRawFnJS(): string {
-      return this._makeFnJS(this.lhs._getRawFnJS(), this.rhs._getRawFnJS())
-    }
-
-    protected _checkTypeOf(lhsRhs: string, wantedType: string): void {
-      var operand: Expression = (<any>this)[lhsRhs];
-      if (!operand.canHaveType(wantedType)) {
-        throw new TypeError(this.op + ' ' + lhsRhs + ' must be of type ' + wantedType);
-      }
-    }
-
-    public _fillRefSubstitutions(typeContext: any, alterations: Alteration[]): any {
-      this.lhs._fillRefSubstitutions(typeContext, alterations);
-      this.rhs._fillRefSubstitutions(typeContext, alterations);
-      return this.type;
-    }
-  }
-
-// =====================================================================================
-// =====================================================================================
-
-  export class NaryExpression extends Expression {
-    static jsToValue(parameters: ExpressionJS): ExpressionValue {
-      var op = parameters.op;
-      var value: ExpressionValue = {
-        op: op
-      };
-      if (Array.isArray(parameters.operands)) {
-        value.operands = parameters.operands.map((operand) => Expression.fromJSLoose(operand));
-      } else {
-        throw new TypeError("must have a operands");
-      }
-
-      return value;
-    }
-
-    public operands: Expression[];
-
-    constructor(parameters: ExpressionValue, dummyObject: Dummy) {
-      super(parameters, dummyObject);
-      this.operands = parameters.operands;
-    }
-
-    public valueOf(): ExpressionValue {
-      var value = super.valueOf();
-      value.operands = this.operands;
-      return value;
-    }
-
-    public toJS(): ExpressionJS {
-      var js = super.toJS();
-      js.operands = this.operands.map((operand) => operand.toJS());
-      return js;
-    }
-
-    public equals(other: NaryExpression): boolean {
-      if (!(super.equals(other) && this.operands.length === other.operands.length)) return false;
-      var thisOperands = this.operands;
-      var otherOperands = other.operands;
-      for (var i = 0; i < thisOperands.length; i++) {
-        if (!thisOperands[i].equals(otherOperands[i])) return false;
-      }
-      return true;
-    }
-
-    public getComplexity(): number {
-      var complexity = 1;
-      var operands = this.operands;
-      for (var i = 0; i < operands.length; i++) {
-        complexity += operands[i].getComplexity();
-      }
-      return complexity;
-    }
-
-    protected _specialSimplify(simpleOperands: Expression[]): Expression {
-      return null;
-    }
-
-    public simplify(): Expression {
-      var simpleOperands: Expression[] = this.operands.map((operand) => operand.simplify());
-
-      var special = this._specialSimplify(simpleOperands);
-      if (special) return special;
-
-      var literalOperands = simpleOperands.filter((operand) => operand.isOp('literal'));
-      var nonLiteralOperands = simpleOperands.filter((operand) => !operand.isOp('literal'));
-      var literalExpression = new LiteralExpression({
-        op: 'literal',
-        value: this._makeFn(literalOperands.map((operand) => operand.getFn()))()
-      });
-
-      if (nonLiteralOperands.length) {
-        nonLiteralOperands.push(literalExpression);
-        var value = this.valueOf();
-        value.operands = nonLiteralOperands;
-        return new (Expression.classMap[this.op])(value);
-      } else {
-        return literalExpression
-      }
-    }
-
-    public getReferences(): string[] {
-      return Array.prototype.concat.apply([], this.operands.map((operand) => operand.getReferences())).sort();
-    }
-
-    public getOperandOfType(type: string): Expression[] {
-      return this.operands.filter((operand) => operand.isOp(type));
-    }
-
-    public substitute(substitutionFn: SubstitutionFn, genDiff: number): Expression {
-      var sub = substitutionFn(this, genDiff);
-      if (sub) return sub;
-      var subOperands = this.operands.map((operand) => operand.substitute(substitutionFn, genDiff));
-      if (this.operands.every((op, i) => op === subOperands[i])) return this;
-      var value = this.valueOf();
-      value.operands = subOperands;
-      return new (Expression.classMap[this.op])(value);
-    }
-
-    protected _makeFn(operandFns: Function[]): Function {
-      throw new Error("should never be called directly");
-    }
-
-    public getFn(): Function {
-      return this._makeFn(this.operands.map((operand) => operand.getFn()));
-    }
-
-    protected _makeFnJS(operandFnJSs: string[]): string {
-      throw new Error("should never be called directly");
-    }
-
-    /* protected */
-    public _getRawFnJS(): string {
-      return this._makeFnJS(this.operands.map((operand) => operand._getRawFnJS()));
-    }
-
-    protected _checkTypeOfOperands(wantedType: string): void {
-      var operands = this.operands;
-      for (var i = 0; i < operands.length; i++) {
-        if (!operands[i].canHaveType(wantedType)) {
-          throw new TypeError(this.op + ' must have an operand of type ' + wantedType + ' at position ' + i);
-        }
-      }
-    }
-
-    public _fillRefSubstitutions(typeContext: any, alterations: Alteration[]): any {
-      var operands = this.operands;
-      for (var i = 0; i < operands.length; i++) {
-        operands[i]._fillRefSubstitutions(typeContext, alterations);
-      }
-      return this.type;
-    }
-  }
-
 }
